Update Dockerfile to facilitate Node.js Build
A dual-stage build process has been implemented in the Dockerfile. The first section is a build stage utilizing a Node.js 20 image to install the necessary node packages and perform a build. This produces a compilation of the codes. In the second stage, a Python 3.11-slim image is used to continue with Python application setup. The built Node files are then copied into this Python runtime environment. This approach optimizes the final image size.