Refactor downloading process and URL routing
The code has been refactored to split the downloading process into two steps - getting the repository information (username and repo name) and downloading the repository. This revision also includes changes to the URL routing to accommodate the separate downloading step. Error handling has been improved with the addition of redirecting and passing of error messages in sessions.