Refactor binary mode detection function in file_utils
The previous function for binary mode detection of file objects was overly complex and hard to maintain. It was reduced and simplified. Support for a wider range of standard Python file-like objects was introduced in the process. The types now being checked include io.BufferedIOBase, bytes, bytearray and memoryview for binary types, and str and io.TextIOBase for text types.