Add dynamic max repo size limit
Introduced a new setting `MAX_REPO_SIZE` to dynamically control the maximum repository size that the system can handle. This maximum size value is now passed as context data to the downloader's template. Furthermore, the file size check in `downloader.js` was updated to handle this dynamic limit instead of a hardcoded size limit.