Add CSS rule for button and clean up html
A CSS rule has been added to the downloader.html template to limit the maximum width of buttons to 250px. Additionally, some formatting and whitespace cleanup has also been done in the same file. There are no functional changes or changes to the visible user interface outside of the button width change.