Add new type definitions and selections count tracking
The new type definitions significantly simplify the handling of selection state changes. The `TreeStateManager` now keeps track of the number of selected nodes and "indeterminate" nodes, which is then used in the updated `selectionChanged` event. Additionally, the `SelectionChangeEventDetail` type has been moved to a separate `eventTypes.ts` file to improve code organisation.