Add function to ensure unique file paths
A function, getUniquePath, has been included to generate unique paths for files. It checks if a given base path has been used and if so, it appends a counter to the path to make it unique. This prevents potential conflicts when processing multiple files with the same path or name. The function is used when creating new ZipPassThrough objects.

This is necessary for browsers that don't support `webkitrelativepath` and thus only give us filenames. In that case we'd get filename collisions when the same-named file existed in more than one directory.